161025_Django_Day13/mysite-project/requirements.txt
coolsms-python-sdk==2.0.1
Django==1.10.2
django-appconf==1.0.2
django-compressor==2.1
django-extensions==1.7.4
django-sass-processor==0.5.1
google-api-python-client==1.5.4
httplib2==0.9.2
libsass==0.11.1
oauth2client==4.0.0
Pillow==3.4.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
rcssmin==1.0.6
requests==2.11.1
rjsmin==1.0.12
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
